rename nfts view to marketplace
