+++ packages/atlas/src/views/viewer/MarketplaceView/index.ts
@@ -1,0 +1,1 @@
+export * from './MarketplaceView'
