--- packages/atlas/src/views/viewer/NftsView/index.ts
@@ -1,1 +1,0 @@
-export * from './NftsView'
