+++ packages/atlas/src/views/viewer/ViewerLayout.tsx
@@ -25,9 +25,9 @@ import { ChannelsView } from './ChannelsView'
 import { DiscoverView } from './DiscoverView'
 import { EditMembershipView } from './EditMembershipView'
 import { HomeView } from './HomeView'
+import { MarketplaceView } from './MarketplaceView'
 import { MemberView } from './MemberView'
 import { NewView } from './NewView'
-import { NftsView } from './NftsView'
 import { NotFoundView } from './NotFoundView'
 import { PopularView } from './PopularView'
 import { SearchView } from './SearchView'
@@ -44,7 +44,7 @@ const viewerRoutes = [
   { path: relativeRoutes.viewer.channel(), element: <ChannelView /> },
   { path: relativeRoutes.viewer.category(), element: <CategoryView /> },
   { path: relativeRoutes.viewer.member(), element: <MemberView /> },
-  { path: relativeRoutes.viewer.nfts(), element: <NftsView /> },
+  { path: relativeRoutes.viewer.marketplace(), element: <MarketplaceView /> },
   ...(atlasConfig.features.ypp.googleConsoleClientId
     ? [{ path: relativeRoutes.viewer.ypp(), element: <YppLandingView /> }]
     : []),
